// src/app/core/enums/theme.ts
export enum Theme {
  darkTheme = 'dark-theme'
}

export enum DarkComponents {
  card = 'dark-card',
  dialog = 'dark-dialog',
  button = 'dark-button',
  toolbar = 'dark-toolbar',
  chip = 'dark-chips',
  dataView = 'dark-dataview',
  selectButton = 'dark-selectButton',
  menuButton = 'dark-menuButton'
}

export enum LightComponents {
  card = 'light-card',
  dialog = 'light-dialog',
  button = 'light-button',
  toolbar = 'light-toolbar',
  chip = 'light-chips',
  dataView = 'light-dataview',
  selectButton = 'light-selectButton',
  menuButton = 'light-menuButton'
}